fixup! GH-140 Replace `Filters` in favor of `Vec<(Filter<'_>, QoS)>'`
