+++ src/encode.rs
@@ -1,134 +1,4 @@
-/// A namespace where messages are published to.
-/// A `Topic` must be valid utf-8 with a minimum length of 1 bytes.
-/// It can't contain the wildcards `#` and `+`.
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct Topic<'a>(&'a str);
-
-impl<'a> Topic<'a> {
-    pub fn new(value: &'a str) -> Result<Self, ValueError> {
-        if value.is_empty() {
-            return Err(ValueError::new("Topic", "must have at least one byte"));
-        }
-
-        verify_utf8(value).unwrap();
-        if value.contains('#') || value.contains('+') {
-            return Err(ValueError::new(
-                "Topic",
-                "contains the wildcard '#' and/or a '+'",
-            ));
-        }
-
-        Ok(Self(value))
-    }
-
-    /// Return the inner string slice.
-    pub fn as_str(&self) -> &str {
-        self.0
-    }
-}
-
-impl<'a> PartialEq<&str> for Topic<'a> {
-    fn eq(&self, other: &&str) -> bool {
-        self.0 == *other
-    }
-}
-
-impl AsRef<str> for Topic<'_> {
-    fn as_ref(&self) -> &str {
-        self.0
-    }
-}
-
-impl std::fmt::Display for Topic<'_> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-// Filter must be at least 1 char long.
-// Must be valid utf-8.
-// Wildcard '#' is the last char of a filter.
-// Wildcard '#' is preceded by nothing or a '/'
-// Wildcard '+' must be preceded by nothing or a '/'
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct Filter<'a>(&'a str);
-
-impl<'a> Filter<'a> {
-    pub fn new(value: &'a str) -> Result<Self, ValueError> {
-        if value.is_empty() {
-            return Err(ValueError::new("Topic", "must have at least one byte"));
-        }
-
-        verify_utf8(value).unwrap();
-        for (index, pattern) in value.match_indices(['+', '#']) {
-            if pattern == "#" {
-                // The '#' can appear only once and it must be the last character of
-                // a filter.
-                if index + 1 != value.len() {
-                    panic!()
-                }
-
-                // If '#' is _not_ the only character in the filter, must be
-                // preceded by a '/'.
-                if index > 0 && value.chars().nth(index - 1).unwrap() != '/' {
-                    panic!()
-                }
-            }
-
-            if pattern == "+" {
-                // If '+' is _not_ the only character in the filter, must be
-                // preceded by a '/'.
-                if index > 0 && value.chars().nth(index - 1).unwrap() != '/' {
-                    panic!()
-                }
-            }
-        }
-
-        Ok(Filter(value))
-    }
-
-    /// Return the inner string slice.
-    pub fn as_str(&self) -> &str {
-        self.0
-    }
-}
-
-#[derive(Debug)]
-pub struct ValueError {
-    field: String,
-    problem: String,
-}
-
-impl ValueError {
-    pub fn new(field: impl Into<String>, problem: impl Into<String>) -> Self {
-        Self {
-            field: field.into(),
-            problem: problem.into(),
-        }
-    }
-}
-
-impl std::error::Error for ValueError {}
-
-impl std::fmt::Display for ValueError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{} is not  valid: {}", self.field, self.problem)
-    }
-}
-
-pub fn verify_utf8(value: &str) -> Result<(), EncodingError> {
-    // 1.5.3 [..] you cannot use a string that would encode to more than 65_535 bytes.
-    if value.len() > 65_535 {
-        return Err(EncodingError::TooLong);
-    }
-
-    // [MQTT-1.5.3-2] A UTF-8 encoded string MUST NOT include an encoding of the null character U+0000.
-    if value.contains('\0') {
-        return Err(EncodingError::IllegalValue);
-    }
-
-    Ok(())
-}
+use crate::verify_utf8;
 
 /// Encode a string as bytes.
 ///
